hemostasis-automation/app/devices/point_of_care.py
# Core logic for Point-of-Care devices
class PointOfCareDevice:
    def __init__(self, device_id):
        self.device_id = device_id

    def get_test_results(self):
        # Simulate fetching test results from a Point-of-Care device
        return {
            "device": "GEM Premier 5000",
            "tests": {
                "Blood Gas": "7.4",
                "Lactate": "1.8"
            },
            "status": "completed"
        }

    def run_calibration(self):
        # Simulate a calibration of the device
        return "calibration successful"

